delt opp filer. lagt til typer og utils.
